requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
python-dotenv==1.0.0
redis==5.0.1
websockets==12.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
